created shared overview of an order component
