+++ projects/storefrontlib/src/cms-components/myaccount/order/order-details/order-details.module.ts
@@ -13,8 +13,11 @@ import {
 import { ConsignmentTrackingComponent } from '../../../../cms-components/myaccount/order/order-details/order-detail-items/consignment-tracking/consignment-tracking.component';
 import { CmsPageGuard } from '../../../../cms-structure/guards/cms-page.guard';
 import { PageLayoutComponent } from '../../../../cms-structure/page/page-layout/page-layout.component';
-import { CardModule } from '../../../../shared/components/card/card.module';
-import { SpinnerModule } from '../../../../shared/components/spinner/spinner.module';
+import {
+  CardModule,
+  OrderOverviewModule,
+  SpinnerModule,
+} from '../../../../shared/index';
 import { CartSharedModule } from '../../../cart/cart-shared/cart-shared.module';
 import { PromotionsModule } from '../../../checkout/components/promotions/promotions.module';
 import { OrderDetailActionsComponent } from './order-detail-actions/order-detail-actions.component';
@@ -45,7 +48,9 @@ const moduleComponents = [
     I18nModule,
     FeaturesConfigModule,
     PromotionsModule,
+    OrderOverviewModule,
     UrlModule,
+    SpinnerModule,
     RouterModule.forChild([
       {
         path: null,
@@ -60,7 +65,6 @@ const moduleComponents = [
         data: { cxRoute: 'orderDetails' },
       },
     ]),
-    SpinnerModule,
   ],
   providers: [
     provideDefaultConfig(<CmsConfig | FeaturesConfig>{
