+++ projects/storefrontlib/src/cms-components/order-confirmation/order-confirmation.module.ts
@@ -7,18 +7,21 @@ import {
   I18nModule,
   provideDefaultConfig,
 } from '@spartacus/core';
-import { CardModule } from '../../shared/components/card/card.module';
+import {
+  CardModule,
+  FormErrorsModule,
+  OrderOverviewModule,
+} from '../../shared/index';
 import { CartSharedModule } from '../cart/cart-shared/cart-shared.module';
+import { PromotionsModule } from '../checkout/components/promotions/promotions.module';
 import { PwaModule } from './../../cms-structure/pwa/pwa.module';
+import { GuestRegisterFormComponent } from './components/guest-register-form/guest-register-form.component';
 import { OrderConfirmationItemsComponent } from './components/order-confirmation-items/order-confirmation-items.component';
 import { OrderConfirmationOverviewComponent } from './components/order-confirmation-overview/order-confirmation-overview.component';
 // tslint:disable-next-line
 import { OrderConfirmationThankYouMessageComponent } from './components/order-confirmation-thank-you-message/order-confirmation-thank-you-message.component';
 import { OrderConfirmationTotalsComponent } from './components/order-confirmation-totals/order-confirmation-totals.component';
-import { GuestRegisterFormComponent } from './components/guest-register-form/guest-register-form.component';
 import { OrderConfirmationGuard } from './guards/order-confirmation.guard';
-import { PromotionsModule } from '../checkout/components/promotions/promotions.module';
-import { FormErrorsModule } from '../../shared/index';
 
 const orderConfirmationComponents = [
   OrderConfirmationItemsComponent,
@@ -39,6 +42,7 @@ const orderConfirmationComponents = [
     ReactiveFormsModule,
     FeaturesConfigModule,
     FormErrorsModule,
+    OrderOverviewModule,
   ],
   providers: [
     provideDefaultConfig(<CmsConfig>{
