+++ projects/storefrontlib/src/shared/components/index.ts
@@ -7,6 +7,7 @@ export * from './item-counter/index';
 export * from './list-navigation/index';
 export * from './media/index';
 export * from './modal/index';
+export * from './order-overview/index';
 export * from './spinner/index';
 export * from './split-view/index';
 export * from './star-rating/index';
