+++ projects/storefrontlib/src/shared/components/order-overview/index.ts
@@ -1,0 +1,2 @@
+export * from './order-overview.component';
+export * from './order-overview.module';
