+++ projects/storefrontstyles/scss/components/misc/_index.scss
@@ -3,6 +3,7 @@
 @import './spinner/index';
 @import './icons';
 @import './table/table';
+@import './order-overview/index';
 
 $misc-components-whitelist: cx-table, cx-icon, cx-spinner, cx-card,
-  cx-pagination !default;
+  cx-pagination, cx-order-overview !default;
