+++ projects/storefrontstyles/scss/components/misc/order-overview/_index.scss
@@ -1,0 +1,1 @@
+@import './order-overview';
